Make $quake and $quake_scope accessible everywhere
Make $quake a constant value so that it can be retrieved by the engine
state instead of the stack, and add $quake_scope to the captures of all
blocks in the source file. This allows us to be able to run commands
like `depends`, etc. inside of functions that have been called inside of
special quake blocks.